feat(output-override): add a way to override the output functions
When unit-testing on devices which don't have terminals it can be useful
to override the output functions.

This provides complete flexibility to the user where and how to output
the data - write it into a shared-memory buffer, send it over a network
port, call a vendor-specific logging API, ... etc.

This commit adds a new API, "cmocka_set_output_callbacks", which takes
two functions, one each for writing standard and error output messages.
These callbacks will be called for almost all of CMocka's output. The
notable exception is the XML output, which is always written to a file
pointer.

WARNING: This commit also removes two public functions from the CMocka
header file, vprint_message and vprint_error. It is unlikely for client
code to invoke these directly, and by removing them this diff could be
kept shorter. It would be trivial to reinstate these functions.

Signed-off-by: Micha WERLE <micha@michaelwerle.com>